requirements.txt
streamlit
torch
transformers
Pillow
python-dotenv
huggingface_hub
